# perf(cli): faster standalone executable determination